Split up humongous file
