+++ src/sentry/testutils/cases/auth/__init__.py
@@ -1,0 +1,12 @@
+from .auth_provider import AuthProviderTestCase
+from .permissions import PermissionTestCase
+from .scim import SCIMAzureTestCase, SCIMTestCase
+from .two_factor import TwoFactorAPITestCase
+
+__all__ = (
+    "AuthProviderTestCase",
+    "PermissionTestCase",
+    "SCIMAzureTestCase",
+    "SCIMTestCase",
+    "TwoFactorAPITestCase",
+)
