+++ src/sentry/testutils/cases/auth/permissions.py
@@ -1,0 +1,85 @@
+from ..base import TestCase
+
+
+class PermissionTestCase(TestCase):
+    def setUp(self):
+        super().setUp()
+        self.owner = self.create_user(is_superuser=False)
+        self.organization = self.create_organization(
+            owner=self.owner, flags=0  # disable default allow_joinleave access
+        )
+        self.team = self.create_team(organization=self.organization)
+
+    def assert_can_access(self, user, path, method="GET", **kwargs):
+        self.login_as(user, superuser=user.is_superuser)
+        resp = getattr(self.client, method.lower())(path, **kwargs)
+        assert resp.status_code >= 200 and resp.status_code < 300
+
+    def assert_cannot_access(self, user, path, method="GET", **kwargs):
+        self.login_as(user, superuser=user.is_superuser)
+        resp = getattr(self.client, method.lower())(path, **kwargs)
+        assert resp.status_code >= 300
+
+    def assert_member_can_access(self, path, **kwargs):
+        return self.assert_role_can_access(path, "member", **kwargs)
+
+    def assert_teamless_member_can_access(self, path, **kwargs):
+        user = self.create_user(is_superuser=False)
+        self.create_member(user=user, organization=self.organization, role="member", teams=[])
+
+        self.assert_can_access(user, path, **kwargs)
+
+    def assert_member_cannot_access(self, path, **kwargs):
+        return self.assert_role_cannot_access(path, "member", **kwargs)
+
+    def assert_manager_cannot_access(self, path, **kwargs):
+        return self.assert_role_cannot_access(path, "manager", **kwargs)
+
+    def assert_teamless_member_cannot_access(self, path, **kwargs):
+        user = self.create_user(is_superuser=False)
+        self.create_member(user=user, organization=self.organization, role="member", teams=[])
+
+        self.assert_cannot_access(user, path, **kwargs)
+
+    def assert_team_admin_can_access(self, path, **kwargs):
+        return self.assert_role_can_access(path, "admin", **kwargs)
+
+    def assert_teamless_admin_can_access(self, path, **kwargs):
+        user = self.create_user(is_superuser=False)
+        self.create_member(user=user, organization=self.organization, role="admin", teams=[])
+
+        self.assert_can_access(user, path, **kwargs)
+
+    def assert_team_admin_cannot_access(self, path, **kwargs):
+        return self.assert_role_cannot_access(path, "admin", **kwargs)
+
+    def assert_teamless_admin_cannot_access(self, path, **kwargs):
+        user = self.create_user(is_superuser=False)
+        self.create_member(user=user, organization=self.organization, role="admin", teams=[])
+
+        self.assert_cannot_access(user, path, **kwargs)
+
+    def assert_team_owner_can_access(self, path, **kwargs):
+        return self.assert_role_can_access(path, "owner", **kwargs)
+
+    def assert_owner_can_access(self, path, **kwargs):
+        return self.assert_role_can_access(path, "owner", **kwargs)
+
+    def assert_owner_cannot_access(self, path, **kwargs):
+        return self.assert_role_cannot_access(path, "owner", **kwargs)
+
+    def assert_non_member_cannot_access(self, path, **kwargs):
+        user = self.create_user(is_superuser=False)
+        self.assert_cannot_access(user, path, **kwargs)
+
+    def assert_role_can_access(self, path, role, **kwargs):
+        user = self.create_user(is_superuser=False)
+        self.create_member(user=user, organization=self.organization, role=role, teams=[self.team])
+
+        self.assert_can_access(user, path, **kwargs)
+
+    def assert_role_cannot_access(self, path, role, **kwargs):
+        user = self.create_user(is_superuser=False)
+        self.create_member(user=user, organization=self.organization, role=role, teams=[self.team])
+
+        self.assert_cannot_access(user, path, **kwargs)
