+++ src/sentry/testutils/cases/integrations/__init__.py
@@ -1,0 +1,13 @@
+from .activity import ActivityTestCase
+from .base import IntegrationTestCase
+from .plugin import PluginTestCase
+from .repository import IntegrationRepositoryTestCase
+from .slack import SlackActivityNotificationTest
+
+__all__ = (
+    "ActivityTestCase",
+    "IntegrationRepositoryTestCase",
+    "IntegrationTestCase",
+    "PluginTestCase",
+    "SlackActivityNotificationTest",
+)
