+++ src/sentry/testutils/cases/snuba/incidents.py
@@ -1,0 +1,34 @@
+from __future__ import annotations
+
+from uuid import uuid4
+
+from django.utils import timezone
+from django.utils.functional import cached_property
+
+from sentry.testutils.helpers.datetime import iso_format
+
+from .base import SnubaTestCase
+
+
+class BaseIncidentsTest(SnubaTestCase):
+    def create_event(self, timestamp, fingerprint=None, user=None):
+        event_id = uuid4().hex
+        if fingerprint is None:
+            fingerprint = event_id
+
+        data = {
+            "event_id": event_id,
+            "fingerprint": [fingerprint],
+            "timestamp": iso_format(timestamp),
+            "type": "error",
+            # This is necessary because event type error should not exist without
+            # an exception being in the payload
+            "exception": [{"type": "Foo"}],
+        }
+        if user:
+            data["user"] = user
+        return self.store_event(data=data, project_id=self.project.id)
+
+    @cached_property
+    def now(self):
+        return timezone.now().replace(minute=0, second=0, microsecond=0)
